+++ cmd/uninstall_unix.go
@@ -1,0 +1,9 @@
+//go:build !windows
+
+package cmd
+
+import "os"
+
+func removeCurrentExecutable(path string) error {
+	return os.Remove(path)
+}